Assistants Stream: Resolved type issue with streaming tool calls
Currently defaulting it to an array,
While the 'delta' is generating the function call all of the data is not there. So its not actually useful, until the `thread.run.requires_action` event is sent. At which point we have the full function call mapped to the correct types anyway.